requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
bcrypt==4.0.1
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
Jinja2==3.1.4
Mako==1.3.3
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
pydentic==0.0.1.dev3
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-stdnum==1.20
PyYAML==6.0.1
redis==5.0.7
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
